added checks for required commands and PWD context

+++ glasp.sh
@@ -2,6 +2,16 @@
 
 ACTION=$1
 
+for COM in clasp git; do
+    command -v $COM >/dev/null || \
+        { echo "Requierd command $COM not found. Aborting." >&2; exit 1; }
+done
+
+if [ ! -f .clasp.json ]; then
+    echo "Did not find '.clasp.json' file. Execute from clasp root directory." >&2
+    exit 1
+fi
+
 if [ x"$ACTION" == x"pull" ]; then
    cp "$PWD/.glaspignore" "$PWD/.claspignore"
    # on pull, also ignore the library files, which are only pushed
